Optimize p4d tests
* Merge nccl test in test_hit_efa
* Run test_hit_efa on all OSes with p4d instances
* Reduce number of p4d instances required by test_hit_efa test by forcing head node to c5.xlarge and removing efa-disabled queue
* Move p4d tests from common.yaml and develop.yaml to p4d.yaml
* Run p4d tests in us-west-2 only
