+++ tests/integration-tests/configs/develop.yaml
@@ -83,22 +83,18 @@ test-suites:
           instances: {{ common.INSTANCES_DEFAULT_ARM }}
           oss: {{ common.OSS_GOVCLOUD_ARM }}
           schedulers: ["slurm"]
+  # TODO: remove p4d test
   multiple_nics:
     test_multiple_nics.py::test_multiple_nics:
       dimensions:
-        - regions: ["us-east-1"]
-          instances: ["p4d.24xlarge"]
-          oss: ["alinux2", "centos8"]
-          schedulers: ["slurm"]
         - regions: ["us-west-2"]
           instances: ["p4d.24xlarge"]
-          oss: ["ubuntu1804", "centos7", "ubuntu2004"]
+          oss: {{ common.OSS_COMMERCIAL_X86 }}
           schedulers: ["slurm"]
-  # nccl tests overwrite the tests in common.yaml
   nccl:
     test_nccl.py::test_nccl:
       dimensions:
-        - regions: ["eu-west-1"]
+        - regions: ["us-west-2"]
           instances: ["p4d.24xlarge"]
           oss: ["alinux2", "ubuntu1804"]
           schedulers: ["slurm"]
