+++ tests/integration-tests/tests/efa/test_efa/test_hit_efa/pcluster.config.ini
@@ -9,8 +9,8 @@ base_os = {{ os }}
 key_name = {{ key_name }}
 vpc_settings = parallelcluster-vpc
 scheduler = {{ scheduler }}
-master_instance_type = {{ instance }}
-queue_settings = efa-enabled,efa-disabled
+master_instance_type = c5.xlarge  # Force c5.xlarge to limit usage of EFA instances to compute nodes only
+queue_settings = efa-enabled
 
 [queue efa-enabled]
 compute_resource_settings = efa-enabled_i1
@@ -20,19 +20,9 @@ enable_efa_gdr = true
 {% endif %}
 placement_group = DYNAMIC
 
-[queue efa-disabled]
-compute_resource_settings = efa-disabled_i1
-enable_efa = false
-placement_group = DYNAMIC
-
 [compute_resource efa-enabled_i1]
 instance_type = {{ instance }}
 min_count = {{ max_queue_size }}
-max_count = {{ max_queue_size }}
-
-[compute_resource efa-disabled_i1]
-instance_type = {{ instance }}
-min_count = 1
 max_count = {{ max_queue_size }}
 
 [vpc parallelcluster-vpc]
