--- tests/integration-tests/tests/nccl/test_nccl/test_nccl/pcluster.config.ini
@@ -1,26 +1,0 @@
-[global]
-cluster_template = default
-
-[aws]
-aws_region_name = {{ region }}
-
-[cluster default]
-base_os = {{ os }}
-key_name = {{ key_name }}
-vpc_settings = parallelcluster-vpc
-scheduler = {{ scheduler }}
-master_instance_type = {{ instance }}
-compute_instance_type = {{ instance }}
-initial_queue_size = 2
-maintain_initial_size = true
-max_queue_size = 10
-enable_efa = compute
-enable_efa_gdr = compute
-placement_group = DYNAMIC
-disable_hyperthreading=true
-
-[vpc parallelcluster-vpc]
-vpc_id = {{ vpc_id }}
-master_subnet_id = {{ public_subnet_id }}
-compute_subnet_id = {{ private_subnet_id }}
-use_public_ips = true
